separated header into h and cpp files -dan
