+++ .gitignore
@@ -1,0 +1,3 @@
+.pio
+.vscode
+platformio.ini
